common: added transport getopt, and support for SO_PEERCRED.
Added infrastructural support for transport getopt. The API is now
more symmteric having both setopt and getopt.

Also added support for retrieving peer credential (SOL_PEERCRED)
over UNIX domain stream sockets using the new getopt infra. While
this is probably the most straightforward abstraction, I'm not sure
if this is the the best/right one... A viable alternative would be
to add support/abstraction for authentication to the transport infra
and plug PEERCRED support into that as one of the backends.