launcher: fix a sh*tload of socket-activation related problems.
Also added a hidden -D option which allows one to specify an
initial startup delay in seconds. This can be used to buy some
time for attaching with a debugger to the launcher daemon when
it is socket-activated (and you have failed to figure out where
why and what goes wrong just by reading the code).