build-sys, common, app: don't assume libuv is packaged separately.
Don't assume/require that libuv is packaged as a separately
installable entity of its own. If we do not find libuv via
pkg-config but NodeJS is present, assume libuv to be part of
NodeJS and adjust the necessary flags accordingly. In this
case also disable all code that would require a standalone
libuv library.

Note that because NodeJS does not use/provide pkg-config files,
we cannot detect/handle if it is installed in a non-standard
location (IOW, headers not in /usr/include/node). Actually,
right now I'm not sure if even cross-compilation for Yocto
works correctly, or if it only seem to work and in cases when
you have NodeJS installed on the build host.