launcher, app: allow non-object JSON event data.
I'm not sure this is a good idea... However, since it was not
obvious for people that we intended to allow only JSON objects
to be attached to events and it was neither documented, nor
trying to do so gave an error, let's support this for now.

If it turns out to be too problematic, we'll switch to allow
only explicitly checked JSON objects.