nodejs, build-sys: dual support for NodeJS 0.10.x/0.12.y.
Cleaned up the NodeJS bindings and reorganized the code to make it
slightly easier to support multiple versions of NodeJS, and added
support for NodeJS 0.12.6.

Updated the build system to allow building against a given version
of NodeJS, in case there are multiple ones installed. Note that as
a side-effect this changed the NodeJS-related command line option
for configure from --enable-nodejs to --with-node[=/prefix/of/node].

Also added support for building the NodeJS bindings without using
npm at all. The old method of building with npm is disabled for now.