build-sys: enable NodeJS-specific CXXFLAGS for quirky NodeJS setups.
Enable NodeJS-specific CXXFLAGS before testing for node.h. This is
necessary, for instance, on Fedora 22 where the basic platform comes
with libuv 1.0, but NodeJS uses 0.10 which is installed to a non-
standard location.

Note that we'll probably have to give up our naively simplistic
--enable-uv/--disable-uv option and replace it with a much more
complex but flexible --with-uv=<name-of-uv-pc-file> approach.
That'd already be flexible enough to handle Fedora 22 and similar
setups.