launcher, app: switch to new msg interface, app listing, bug fixes.
Switched the launcher and the app lib to the new msg interface, re-
added initial listing of running apps, fixed a number of various bugs.

Hmm... I should have been more disciplined and committed this in several,
smaller, self-contained pieces. Too late now... I guess. Well, at least
we can use this as a demo commit that should not be to showcase how patches
should not be committed.