utils: first shot at manifest-based path and security type mapping.
This patch adds support for mapping paths to applications
and security-manager path types using the manifest. This
is still work in progress.

The patch adds support for reverse-mapping file paths in
the package to application/security-type tuples. This is
needed by the installer for registering the paths to the
correct application with the correct type so they can be
properly labeled by the security-framework.

There is support for mapping files by explicit full path,
by shell-like globbing patterns, by regexp patterns, and
as a fallback implicitly by being installed under the same
directory hierarchy where the actual application is put.
The file permission to security type guessing logic on the
fallback path needs to be reviewed by someone who has way
better understanding of the semantics of the various
security file types.

Also, the regexp based matching is currently incomplete.
We should either guarantee checking regexps in order of
declaration in the manifest (might be difficult if json-c
returns keys for objects in a unpredictable order), or
we should have longest match wins semantics. Needs a bit
more work...