utils: fix hash-based pathmap file lookup.
pathmap_file_lookup should return -1 to indicate failure if
an entry was not found (IOW if the hash-table gave NULL for
the entry). To make sure NULL is an invalid entry (and not
the entry for 0:0) we always insert path-type + 1. Hence on
the lookup path we need to return entry - 1.