app: argh... don't test for NULL-ness of an uninitialized variable.
Check the correct variable for NULL. I hate that gcc does not
do liveness analysis for variables with -O0 so it cannot give
a warning about the usage of uninitialized variables. It has
bitten me lately too many times. Have to make it  ahabit to do
an optimized compilation as the last check always before doing
a commit even for the smallest change.