package-manager: added support for basic package (signature) verification.
Added support for verifying package signatures. Currently there is
a separate mode for just doing signature verification and also a
flag for turning verifucation on during installation. Once all
packages are expected to be singed, we'll flip the semantics of
the latter around and will do signature verification by default.