service-generator: embrionic systemd service generator.
Added an embrionic version of a systemd service generator. It
scans a well-known directory (/apps by default) for installed
applications and generates a systemd service file for every
application (manifest) found.

Manifests are JSON files that describe the application (provider,
description), and how the application should be run. Currently a
bare minimum of options is supported, limited to user/group ID,
environment variables, whether the appliction should be run inside
a container, a few container options, and the command to start the
application. There are quite a number of things left to add...

Manifest key handlers are dynamically registered so in principle
it should be relatively easy to add support for new manifest keys
or features. In practice this is only true for self contained
manifest keys, those which have no effect on and which are not
affected by the presense/absence/value of other manifest keys.

New container type handlers can also be registered dynamically so
in principle is should be possible to add new container types as
well. In practice because the same cross-dependency is true for
containers/keys as for just keys this probably is/will not be as
easy as it might sound, unless we're significantly improving things.

To see the practical problems related to the cross-dependency
mentioned above consider, for instance, the very simple case of
running the application with the right user ID. Depending on
whether the application is run outside or inside a container, the
user ID should be either added as a simple systemd directive, or
typically as a command-line argument/option to the command boot-
strapping the container for the application. Moreover, what the
user ID should be might not be known yet by the time we process
the directive/key specifying the command line or the container.

To allow for taking such cross-effects into consideration the
service file for each application is generated in two phases. The
first phase only collects and generates data about the application.
Second phase generates/writes the service file based on the data
collected. While this helps a bit with the cross-dependency problem
it is not enough to eliminate it altogether as the key/container
handlers still need to be aware of the data of each other. This is
currently rather ugly and there is plenty of room for improvement.