service-generator: systemd-nspawn patches for service generator.
Added two systemd-nspawn patches for our systemd service generator.
For both patches we have versions against v222, v225 and v228.

The first patch delays the check for the binary/OS-tree only after
all the requested bind- and overlay-mounts have been put in place.
This is necessary because it might be (and in our case it's always)
the case that the binary to execute inside the container is pulled
in by one of the container-specific mounts.

The second patch adds support for automatically starting a DHCP
client inside a (non-system mode) container without --share-system
if private networking is enabled. The client is started before the
binary to run inside the container. We might need to generalize this
patch further to allow us running any number of specified binaries
to prepare the container for running the main binary.