jmpl: initial/proof of concept JSON template library.
As of now, this can be considered an initial proof of concept
implementation of JSON template processing library. The library
produces text (file) from a text template (file) and JSON data.

The templates are 'scriptable'. The template scripting language
can reference/evaluate JSON data to alter the output from the
template. There are simple variable substitution, if-then-else,
and foreach constructs supported by the scriptable templates.

The main reasons this can only be considered a proof-of-concept:

  - the lexical analyser and parser are quite a mess (and limited)
  - no sane error or diagnostic messages are collected or produced
    about parsing or evaluation errors
  - the whitespace-handling strategy is broken, occasionally omitting
    or producing extra whitespace; the strategy/approach has been
    evolving during prototyping quite a bit and the current behavior
    (mostly due to the crappy lexical analyser/parser) is not fully
    the final one
  - the accepted grammar is slightly inconsistent
  - expression parsing is limited, no operator-precedence, can only
    deal with fully parenthesized binary operators
  - type support is limited/a bit vague (explicit support mostly for
    strings, implicit support also for integers, doubles, and booleans)
  - and probably others I've already forgotten about...