WiP: more work on the parser, initial template evaluator.
Okay, as of now, maybe this can be considered a proof of concept.
It is pretty hackish, quite a mess, and also quite limited in an
awful lot of places. I want now to first proceed with this until
the whole chain is in an initial working shape then take a step
back and do a massive refactoring/cleanup. In practice this will
probably pretty much mean reworking the whole thing. We'll also
need to address the worst omissions and the most painful problem
spots (which there might be plenty of).

manifests + templates -> generator -> systemd service files is
the initial chain we need to get working first. The biggest /
ugliest problem spots that we will need to address are expected
to be (in no particular order):

 - no (fully implemented) strategy for exact whitespace control
 - the parser
    o no proper error handling or messages
    o expression parsing is limited, no operator-precedence,
      the half-arsed parser can only deal with fully parenthesized
      logical and (limited) relational expressions
 - supported types (explicit support mostly for strings, implicit
   support also for integers, doubles, and booleans)

For a starter...

Maybe the syntax is complex enough to rewrite the parser with bison
and perhaps the lexical analyzer with flex, although the configurable
keyword delimiters might not make this straightforward or worth the
effort.