service-generator: initial implementation of template-based generation.
This patch changes our whole approach for generating service files.
Instead of the whole logic being (hard)coded in C and some limited
flexibility provided by registrable key handlers, we generate files
based on scriptable templates. The real heavy lifting (the template-
processing functionality) is handled by our JSON template library,
jmpl.

For every application manifest found, the service generator feeds
the data from the manifest, a small amount of additional data, and
the service template to the template library which then generates
the final application-specific systemd service file... if everything
goes well.

Ideally with this patch in place, it is now possible to accomodate
necessary changes both to the generated service files and to the
application manifests without changes to the service generator code
itself.

One part still missing is a full implementation of a JSON-to-JSON
preprocessing phase which would allow one to make arbitrary changes
to the mannifest before it is used for actual service generation.
There is some initial code but the plugin/DSO loading bits are not
in place and the whole shebang has never been really tested. The
idea behind the preprocessing phase is that it could be used, for
instance, for presenting a simplified manifest interface towards
application developers, for doing arbitrary value-mapping in the
manifest, or for providing seamless backward-compatibility when the
manifest format needs to change.