jmpl: added support for loopchecks (?first, !first, ?last, !last).
Hacked in a construct to allow for checking whether one is in the
first, non-first, last, or non-last iteration of a foreach loop and
evaluate an (if-else like) true- or false-branch accordingly.

Also adjusted slightly the rules for scanning ids. id scanning is now
terminated at a directive if this precedes the terminating whitespace.
Finally any trailing ',' and ';' are excluded from the id and pushed
back to the input stream. We probably should push back all characters
not allowed in an id instead. This is merely a hack to allow these
operators to be used within a foreach loop to generate comma- and
semicolon-separated lists.