jmpl: added support for trailcheck blocks (?trail:, !trail:).
Hacked in one more construct to allow for checking the trailing
part of the output produced so far against a string or a regexp,
and execute an (if-else like) true- or false-branch accordingly.
The regexp-matching code paths are not implemented yet...

Also removed a number of #if'd code blocks added during the
leading tabulation filtering cleanup.