common: added iot_json_load_file for loading a JSON object.
Loading a single JSON object from a file was getting used
in multiple places. Hence it is now provided by the common
JSON interface. This version actively checks that the file
has only a single object and refuses to load anything other-
wise. We should provide a more flexible or versatile set of
functions for loading objects...