service-generator: added a group whitelisting preprocessor.
Added a simple manifest preprocessor for filtering out groups
that are not explicitly whitelisted (or do not exist). Groups
can be matched by explicit full group names or simple name
prefixes ('prefix*'). An undefined whitelist disables white-
list checks altogether.

Whitelisted groups are specified as an array of strings assigned
to the key 'GroupWhitelist' in the config file. This might still
change...