smpl: initial commit of our revised template library.
This commit adds an initial/embrionic version of our revised
scriptable template library. It's at the level of feature parity
with the original prototype implementation. It supports similar
if-else branches, foreach loops, macro and variable substitution
in addition to file inclusion, block- and line-comments.

This library can produce diagnostic error messages for template
parsing (loadtime) and evaluation (runtime) failures. The error
handling bits and the produced messages need improvements to be
truely helpful in identifying the problems.

An attempt has been made to structure the code (and especially
the parser) slightly better, with room to spare for adding a few
useful features. Some of these include switch-case constructs,
parametrizable (function-like) macros, and potentially extending
the library with externally registrable macro-like functions.

There are a few things left to be decided about the syntax. One
is, for instance, whether macros should live in the same name-
space as (loop) variables. The alternative would be to switch
macro evaluation over to the invoke-construct we have support for
in the parser. That would keep macro evaluation syntactically very
close to variable substitution. However by requiring one to denote
explicitly macro references as such it would effectively put macros
(or callables) and variables in different namespaces.