smpl: support calling functions registered from C.
Added support for registering functions from C that can then be
called from processed templates. Functions can be registered for
a particular template or globally for all templates. A function
call from a template results in a call of the handler registered
for the function. Functions can receive an arbitrary number of
SMPL values as arguments and return an SMPL value (set to unset
in case no value is returned).

Significant uglification warning:
Ouch... the changes related to value and expression handling to
properly free dynamically allocated data returned from function
handlers turned out to be truely horrible. This really needs to
be reconsidered... It'd be probably better to just blindly ref-
count all values always unconditionally. Or alternative to change
the lifecycle of values passed back from functions so they outlive
template processing, then require the function handlers to track
their own allocations, and once we're done with the template call
a (registered) handler to free all allocated values. Frankly, I
think almost anything would be an improvement...