smpl: fix a cleanup path crash, allow setting symbols to unset value.
Fix a crash when freeing macro or function call expressions. Allow
symbol table values to be set to the 'unset' value to let calling
functions or macros with an uset value mask previously set values
with a conflicting name in the calling scope. Also added dumping
for macro and function references (calls).