service-generator: added update mode (-u).
In update mode, we skip generating service files which already
exist and are newer than their input manifest. Since we consider
a file newer if it has a more recent modification timestamp,
probably we should also set the service timestamp to the
manifest timestamp if we decide to (re)generate it.