service-generator: first shot at firewall addon generation.
Handle hole-punching in the firewall by generating an addon
service that the applications service depends on. This is
necessary because explicitly started (as opposed to socket-
activated) services that get started by systemd as a non-
privileged user are unable to punch holes in the firewall
using pre-/post-exec commands as those are executed after
systemd has dropped privileges.

I'll try to generalize the current idea further. Currently
this is specific to firewall-service generation but it should
be possible to generalize this further so that a service file
can request an arbitrary number of addon generation. Now part
of the logic (related to what template to use and where to
generate the addon files) is in the service generator. The
template simply indicates it needs a firewall addon by calling
back to the generator via a registered function.

This mechanism should be generalized so that the template can
request a particular named addon, specifying an optional
additional input data file, an input template, and an output
file, or token used to determine the exact output location.
This way we could handle similar needs in the future purely
on the template side without having to touch the generator
code itself. And probably this should eventually be lifted
over to the template library instead of keeping it in the
generator.