smpl, service-generator: generalized template addon support.
Rolled a generalized version of the original firewall template
idea and moved support for it into the template library itself.
With this in place a template can now request during evaluation
further 'addons' (usually additinal templates) to be processed.

Requesting an addon from a template happens by calling a builtin
function (REQUEST-ADDON). An addon request needs to provide an
addon name, and can provide a template name and an output des-
tination. At (main) template creation time, one can provide an
addon notification/verdict callback. If given this will be
called whenever an addon is requested during template evaluation.

The notification callback can abort, reject, or accept an addon
by returning 0, < 0, or > 0. Aborting results in a template
evaluation error. Rejection does not add the addon to the main
template but does not abort evaluation. Accepting will add the
addon to the main template and continue evaluation. If accepted,
the callback should also fill in any potentially missing details
(template name, output destination). It can also override the
given details at will.

Silently rejecting an addon can be (mis)used to trigger further
processing outside of the jurisdiction of the template library.
Autostart is a prime example of how this can be used: when the
autostart 'addon' is requested, the notification callback simply
sets a flag in the associated service structure and rejects the
addon. After template processing, the service generator loops thru
all services and symlinks it into the appropriate target if the
autostart flag is set.

The currently provided functionality is still a bit incomplete.
Rolling back/cleaning up if writing one of the necessary addons
fails is not properly done. Also, we'll need an API for looping
through all the addons (and their content) of the main template
(for instance in case someone needs to take care of writing out
the result themselves, or if the generated result is not supposed
to be written out at all).