smpl: fix calling macros with dynamic values (from functions).
Argh... we'll need to clean up the runtime evaluation model. We
should fix at least the following: 1) not being able to pass
the result of macro evaluation as arguments to functions and
macros, 2) not being able to do even the simples shell-like
concatenation (foo-{var1}-bar-{var2}) in argument lists (we need
an explicit CONCAT-function hack for that now), 3) we should
probably turn values into refcounted entities to get rid of the
current fragile value->dynamic braindamage.