smpl: shunting-yard kludge for nested macro- and function calls.
Added an arglist-guard kludge to avoid value_push overeagerly
eating arguments of outer function calls. Without this kludge
constructs like

  {FOO}({BAR}(1, 2, 3), {BAR}(4, 5, 6))

was incorrectly parsed with the reduction of the second nested
call to BAR eating the already reduced first one, instead of
letting the reduction of foo take it.

Hmm... shaky as it goes. This parser is not something to write
home about. The cleanup will be a rewrite, probably using flex
and bison.