smpl: (crude and a bit ugly) support for vararg macros.
Added support for macros with a variable number of arguments.
The syntax is gccish (args...) and the semantics is python-
like (*args).

If you call this macro (assuming your template starts as { } .)

    {macro} VARARG(a, b, c, d, e...) {do}
         {a}:{b}:{c}:{d}
    .    {foreach} arg {in} {e} {do}
    ....     vararg {arg}
    .    {end}
         {e[0]}-{e[1]}-{e[2]}-{e[3]}-{e[100]}
    {end}

with these arguments

    {VARARG}(1, 2, 3, 4, 5, 6, 7, 'a', 'one', 'two', 'three')

you will get the following as output

     1:2:3:4
     vararg 5
     vararg 6
     vararg 7
     vararg a
     vararg one
     vararg two
     vararg three
     5-6-7-a-

The current implementation will cause small memory leaks if
you dereference the vararg macros with the []-indexing notation
and some of the indexed values are dynamic (ie. a non-static
string, JSON-object, or array, returned as a value from a function
call). Try to avoid such complex setups for now.