smpl: support adding auxilary data for addon evaluation.
You can now pass auxilary data for addon evaluation by passing
an argument 'data:<name>' followed by the desired value for
name. For instance

{REQUEST-ADDON}('name:foo', 'destination:foo.file',
                'data:foo', 'bar', 'data:bar', 3.141,
                'data:foobar', 'xyzzy', 'data:meaningoflife', 42)

will set an additional member of the global data passed to the addon
foo:

  'addon': {
      'foo': 'bar',
      'bar': 3.141,
      'foobar': 'xyzzy',
      'meaningoflife': 42
  }

You can also set JSON object and array members, but at the moment
you cannot construct them by direct scripting in the template. The
only way to do it is to dereference a member of the global data,
or to call a function that returns a JSON object or an ARRAY and
use a call to such a function as an argument to REQUEST-ADDON.