iot-service-generator: support reading self manifest, starting services.
Added embrionic support for running/starting as an agent inside a
container. When running in this mode, we use a different template
to generate a 'scriptlet' for ourself, containing a set of simple
instructions we interpret and carry out. Currently the supported
set of commands is

  - exec (execute the given command)
  - setuser (set user ID)
  - and setgroups (set supplementary groups)

We'll use this (in non-system containers) to start up all the
requested processes in the manifest for rendering the service
/ application.

The actual fork/exec code is really primitive. There is nothing
even remotely fancy there, no support for process monitoring or
restart, we also don't stick around waiting for children even if
we're not run as pid 2 (in shared system mode), no double-fork
to reparent to the systemd stub init if we're running as pid 2.
This is work in progress...