WiP: smpl: hackish support for writing out just addons.
We need to better thinks through the current functionality for
writing out results/addon results. It is hackish at best and
too infelxible at worst.

One option would be to a mechanism (smpl_process_result) for
looping through all results (main template and addon) invoking
a callback for each, pssing the generated result and any other
parameters (name, used template, requested destination) as
arguments. The callback is then free to deal with the data as
it best sees fit. Additionally we can provide a simple function
for writing out a generated result to a given destination.