smpl: initial support for callback-based result processing.
Added preliminary support for callback-based result processing.
smpl_process_result takes a result buffer and a callback. It
loops through the main result buffer and any potential addons
it may container, invoking the callback for each and every one.
The is free to do wahtever it sees fit with the result and can
report back failure (processing aborted), success, a stolen
result buffer (gets disassociated from the template/addon), or
request to free the the main result buffer (along with any
associated addon result) or any of the addons.