smpl: added builtin COUNTER function.
Added a builtin counter function. Counters are identified by
a string name. Calls with a single name argument return the current
value assigned to the named counter. Calls with a name and an
integer argument increment the named counter by the given integer
value and return the updated value. All counters are initialized
to 0.